[Refactor] 컨트롤러 및 서비스 분리 - #28
